requirements.txt
crewai==0.30.11
crewai-tools==0.2.6
python-dotenv==1.0.1
streamlit==1.34.0
langchain
langchain-openai
langchain-groq==0.1.4
sec-api==1.0.18
unstructured==0.14.2
agentops==0.2.3
